testing/web-platform/tests/html/browsers/history/the-history-interface/003.html

<!doctype html>
<html>
	<head>
		<title>Synthetic popstate events</title>
		<script type="text/javascript" src="/resources/testharness.js"></script>
		<script type="text/javascript" src="/resources/testharnessreport.js"></script>
		<script type="text/javascript">
var PSEvent, globaldata;
test(function () {
	assert_true( !!window.PopStateEvent );
}, 'PopStateEvent should be exposed for prototyping');
test(function () {
	PSEvent = document.createEvent('PopStateEvent');
	assert_true( !!PSEvent.initPopStateEvent, 'browser created a generic event instead of a PopStateEvent' );
}, 'document.createEvent should create a PopStateEvent');
test(function () {
	PSEvent.initPopStateEvent('popstate',false,true,{testdata:true});
}, 'initPopStateEvent should initialise the event');
test(function () {
	assert_false(PSEvent.bubbles);
	assert_true(PSEvent.cancelable);
	assert_equals(PSEvent.type,'popstate');
	assert_true( PSEvent.state && PSEvent.state.testdata );
}, 'initPopStateEvent properties should hold the requested values');
test(function () {
	assert_true(PSEvent instanceof window.PopStateEvent);
}, 'the resulting object should be instanceof PopStateEvent');
test(function () {
	assert_true(PSEvent instanceof Event);
}, 'the resulting object should be instanceof Event');
window.addEventListener('popstate',function (e) { globaldata = e.state; },false);
test(function () {
	window.dispatchEvent(PSEvent);
}, 'initPopStateEvent should be dispatchable on the window object');
test(function () {
	assert_true(!!globaldata,'state data was lost');
	assert_true(globaldata.testdata,'state data was corrupted');
}, 'state data should be retained correctly when dispatching the event');
		</script>
	</head>
	<body>

		<noscript><p>Enable JavaScript and reload</p></noscript>
		<div id="log"></div>

	</body>
</html>
